refactor: Move SendCoinsRecipient in own header
