+++ src/qt/guiutil.cpp
@@ -7,6 +7,7 @@
 #include <qt/bitcoinaddressvalidator.h>
 #include <qt/bitcoinunits.h>
 #include <qt/qvalidatedlineedit.h>
+#include <qt/sendcoinsrecipient.h>
 #include <qt/walletmodel.h>
 
 #include <base58.h>
