+++ src/qt/paymentserver.h
@@ -36,6 +36,7 @@
 #include <config/bitcoin-config.h>
 #endif
 
+#include <qt/sendcoinsrecipient.h>
 #include <qt/walletmodel.h>
 
 #include <QObject>
