+++ src/qt/receiverequestdialog.h
@@ -5,6 +5,7 @@
 #ifndef BITCOIN_QT_RECEIVEREQUESTDIALOG_H
 #define BITCOIN_QT_RECEIVEREQUESTDIALOG_H
 
+#include <qt/sendcoinsrecipient.h>
 #include <qt/walletmodel.h>
 
 #include <QDialog>
