+++ src/qt/recentrequeststablemodel.h
@@ -5,6 +5,7 @@
 #ifndef BITCOIN_QT_RECENTREQUESTSTABLEMODEL_H
 #define BITCOIN_QT_RECENTREQUESTSTABLEMODEL_H
 
+#include <qt/sendcoinsrecipient.h>
 #include <qt/walletmodel.h>
 
 #include <QAbstractTableModel>
